feat(server): Export config vars directly, not inside an {} 🚀
This allows us to have auto-imports, auto-completions for them and is
just overall better.

Using a default export is not an option, since you get no
auto-completions. Also, it just sucks in general.

Another benefit is that we don't have to wrap them inside some object,
because if you wanted to set some variable based on another, you'd have
to put it outside the object, create the object variable based on it and
then put both of'em inside the object. We used to have this and it
sucks. Off we go!

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>