feat(scraper): Retry on failed scrapes! ⛑
This ensures that we always successfully scrape 100% of the profiles.

Previously, there might've been too many network requests to the website
we're scraping from, and we'd get a 503 error of "service unavailable".

It would result in some students NOT having their profiles available
in our app.
This is exactly the scenario we're handling - we retry later if a
request failed.

Let's goooo!

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>