feat(scraper): Create `memoize` util...
...that we used a lot, but didn't end up using in the end.

Mainly because it caused a lot of confusion
due to the very own nature of memoization itself,
and we did not need that.

But, perhaps in the future we will,
so we're keeping it here just for a sec before we revert the commit
to still be able to find it later.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>