perf(scraper): (BIG) Do NOT populate students with lessons 💣
Instead, populate the lessons once needed.

We currently had 224 students.
This update reduced the database's size from 23MB to 572KB - 41 times! 🔥

It'll be interesting to see if this is worth it, since we'll now have
more overhead for each API request, since we'll have to populate
students with their lessons.
Caching should help for multiple identical requests -- TODO in the future.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>