feat(scraper): Rewrite `extractLessons` to allow individual parsers
While scraping a `Student` or a `Class` is mostly similar,
the details actually vary quite a bit,
mostly because the website we scrape from is a mess!
(that's why we started this project in the first place^^)

This allows us to have individual parsers to extract lessons while still
re-using as much code code as effective.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>