fix(scraper): use proper parsing for lesson attribute entries instead of hard-coded indices
the previous parsing was unreliable, it was falling short often because of
the upstream's whims (which is another topic).

Thus, this change makes it way more robust - we're checking each
individual text entry of the lesson and analysing it if it matches
the given regex.

Results drastically improved. Code was also made simplier - we no longer
need multiple separate parsers - we just use the main one since it
handles all cases (that I could think of) instead of trying to adjust to
the impossible mess that the upstream is.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>