preparation for deploy-thru-remote - get rid of `:nobuild`, instead create `:no-finish` script
original motivation for creating `:nobuild` in [1] were memory issues,
but those have been solved by simply introducing a swap file.

so that's no longer needed.
now, we're adding `:no-finish`, because we need to do everything on the remote:
get latest source, build the docker image, push to container
registry, but not swap the running docker container to a newer image --
the last part will be done on the local machine, because it has the
ARCHIVE_DEPLOY_KEY, thus making it easier & mroe seamless.

recall that the reason why we need to build the docker image on the remote,
as opposed to doing everything locally, is the architecture mismatch,
namely, arm64 (local M1 mac) and x86_64 (remote ubuntu).

[1] cc948d5a05b9cf2294183f4fefa35909ce08e7da